Move targeting bits properties to MarvinToolsClass
